fraud_generator/requirements.txt
pandas>=2.0.0
numpy>=1.24.0
faker>=19.0.0
requests>=2.31.0
python-dotenv>=1.0.0
jsonschema>=4.19.0
colorlog>=6.7.0
ollama>=0.1.0
